Some progress on the pigeon carrier
Not sure what all this one has to do so I'll ask Ms. Healey but for now it takes a message, then shows the bird flying, then prints back the message. Also I'm not sure if the functions of the class are being called properly at the end or something but thats an easy fix and the code part should be fine. If anyone wants to make it compile before I get to it, that could be cool